# apps/api/openapi.admin.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "VoucherShyft API",
    "version": "v1",
    "description": "Internal. No compatibility guarantees."
  },
  "components": {
    "schemas": {}
  },
  "paths": {
    "/admin/tenants": {
      "get": {
        "tags": [
          "admin"
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "tenant_id": {
                            "type": "string"
                          },
                          "host": {
                            "type": "string"
                          },
                          "tenant_slug": {
                            "type": "string",
                            "nullable": true
                          },
                          "status": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "tenant_id",
                          "host",
                          "status"
                        ]
                      }
                    },
                    "correlation_id": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success",
                    "data",
                    "correlation_id"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ]
                    },
                    "correlation_id": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success",
                    "error",
                    "correlation_id"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ]
                    },
                    "correlation_id": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success",
                    "error",
                    "correlation_id"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tenant_id": {
                    "type": "string"
                  },
                  "host": {
                    "type": "string"
                  },
                  "tenant_slug": {
                    "type": "string",
                    "nullable": true
                  },
                  "status": {
                    "type": "string",
                    "nullable": true
                  }
                },
                "required": [
                  "tenant_id",
                  "host"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "tenant_id": {
                          "type": "string"
                        },
                        "host": {
                          "type": "string"
                        },
                        "tenant_slug": {
                          "type": "string",
                          "nullable": true
                        },
                        "status": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "tenant_id",
                        "host",
                        "status"
                      ]
                    },
                    "correlation_id": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success",
                    "data",
                    "correlation_id"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ]
                    },
                    "correlation_id": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success",
                    "error",
                    "correlation_id"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ]
                    },
                    "correlation_id": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success",
                    "error",
                    "correlation_id"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/admin/tenants/{tenant_id}": {
      "patch": {
        "tags": [
          "admin"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "host": {
                    "type": "string",
                    "nullable": true
                  },
                  "tenant_slug": {
                    "type": "string",
                    "nullable": true
                  },
                  "status": {
                    "type": "string",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "tenant_id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "tenant_id": {
                          "type": "string"
                        },
                        "host": {
                          "type": "string"
                        },
                        "tenant_slug": {
                          "type": "string",
                          "nullable": true
                        },
                        "status": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "tenant_id",
                        "host",
                        "status"
                      ]
                    },
                    "correlation_id": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success",
                    "data",
                    "correlation_id"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/admin/tenant-apps": {
      "post": {
        "tags": [
          "admin"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tenant_id": {
                    "type": "string"
                  },
                  "app_key": {
                    "type": "string"
                  },
                  "enabled": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "tenant_id",
                  "app_key",
                  "enabled"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "tenant_id": {
                          "type": "string"
                        },
                        "app_key": {
                          "type": "string"
                        },
                        "enabled": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "tenant_id",
                        "app_key",
                        "enabled"
                      ]
                    },
                    "correlation_id": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success",
                    "data",
                    "correlation_id"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ]
                    },
                    "correlation_id": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success",
                    "error",
                    "correlation_id"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "enum": [
                        false
                      ]
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ]
                    },
                    "correlation_id": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success",
                    "error",
                    "correlation_id"
                  ]
                }
              }
            }
          }
        }
      }
    }
  }
}